+++ src/controller/index.js
@@ -1,9 +1,11 @@
-const express = require('express');
+const express = require("express");
 
-const pictures = require('./picture');
+const login = require("./login");
+const pictures = require("./picture");
 
 const route = express.Router();
 
 route.use(pictures);
+route.use(login);
 
 module.exports = route;
