+++ src/controller/login/index.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const postLogin = require("./postLogin");
+const route = express.Router();
+
+route.post("/login", postLogin);
+
+module.exports = route;
